Refactor imports and fix some weird eslint warning
